17th video sublevel for category

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -9,6 +9,21 @@ use Illuminate\Support\Facades\Storage;
 
 class CategoryController extends Controller
 {
+
+    protected $appends = [
+        'getParentsTree'
+    ];
+
+    public static function getParentsTree ($category, $title){
+        if ($category->parent_id == 1)
+        {
+            return $title;
+        }
+        $parent = Category::find($category->parent_id);
+        $title = $parent->title.' > '. $title;
+        return CategoryController::getParentsTree ($parent, $title);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -30,7 +45,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('admin.category.create');
+        $data = Category::all();
+        return view('admin.category.create', [
+            'data' => $data
+        ]);
     }
 
     /**
@@ -42,7 +60,7 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         $data = new Category;
-        $data->parent_id = 1;
+        $data->parent_id = $request->parent_id;
         $data->title = $request->title;
         $data->keywords = $request->keywords;
         $data->description = $request->description;
@@ -79,8 +97,10 @@ class CategoryController extends Controller
     {
         //
         $data = Category::find($id);
+        $datalist = Category::all();
         return view('admin.category.edit', [
-            'data' => $data
+            'data' => $data,
+            'datalist' => $datalist
         ]);
     }
 
@@ -95,7 +115,7 @@ class CategoryController extends Controller
     {
         //
         $data = Category::find($id);
-        $data->parent_id = 1;
+        $data->parent_id = $request->parent_id;
         $data->title = $request->title;
         $data->keywords = $request->keywords;
         $data->description = $request->description;
@@ -117,7 +137,9 @@ class CategoryController extends Controller
     {
         //
         $data = Category::find($id);
-        Storage::delete($data->image);
+        if ($data->image){
+            Storage::delete($data->image);
+        }
         $data->delete();
         return redirect('admin/category');
